+++ src/main/java/personal/delivery/cart/service/CartMenuService.java
@@ -5,6 +5,7 @@ import personal.delivery.cart.dto.CartMenuResponseDto;
 import personal.delivery.cart.entity.Cart;
 import personal.delivery.cart.entity.CartMenu;
 import personal.delivery.menu.Menu;
+import personal.delivery.order.dto.OrderResponseDto;
 
 import java.util.List;
 
@@ -17,4 +18,6 @@ public interface CartMenuService {
     List<CartMenuResponseDto> getCartMenuList();
 
     CartMenuResponseDto deleteCartMenu(CartMenuDto cartMenuDto) throws Exception;
+
+    OrderResponseDto orderCartMenu(CartMenuDto cartMenuDto);
 }
